faster code: unsigned int

+++ src/timer.c
@@ -26,7 +26,7 @@ double timer_reset(Timer *timer) {
 
   secs = (double)(stop.tv_usec - timer->start.tv_usec) / 1000000 +
          (double)(stop.tv_sec - timer->start.tv_sec);
-  per_secs = ((float)timer->counter) / secs;
+  per_secs = (double)timer->counter / secs;
 
   timer->start = stop;
   timer->counter = 0;
